pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "milestone1"
version = "0.1.0"
description = "AI-powered restaurant recommendations (milestone 1)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "python-dotenv>=1.0.0",
  "datasets>=2.16.0",
  "httpx>=0.27.0",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]
streamlit = [
  "streamlit>=1.28.0",
]

[project.scripts]
milestone1 = "milestone1.cli:main"
milestone1-api = "milestone1.phase6_api.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
  "integration: hits Hugging Face Hub (set RUN_HF_INTEGRATION=1)",
]
